+++ tests/dto_factory/test_dto_factory_model_conversion.py
@@ -6,9 +6,9 @@ from typing import Any, Callable, Dict, List
 import pytest
 from pydantic_factories import ModelFactory
 
+from starlite.contrib.tortoise_orm import TortoiseORMPlugin
 from starlite.dto import DTOFactory
 from starlite.exceptions import ImproperlyConfiguredException
-from starlite.plugins.tortoise_orm import TortoiseORMPlugin
 from tests import Person, TypedDictPerson, VanillaDataClassPerson
 from tests.plugins.tortoise_orm import Tournament
