+++ tests/plugins/tortoise_orm/test_tortoise_orm_plugin_integration.py
@@ -1,4 +1,4 @@
-from starlite.plugins.tortoise_orm import TortoiseORMPlugin
+from starlite.contrib.tortoise_orm import TortoiseORMPlugin
 from starlite.status_codes import HTTP_200_OK, HTTP_201_CREATED
 from starlite.testing import create_test_client
 from tests.plugins.tortoise_orm import (
